event_api/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
mysqlclient==2.1.1
python-decouple==3.7
django-environ==0.9.0
djangorestframework-simplejwt==5.2.2
django_filters
django-cron
icalendar